Really Big refactor, services moved, useless code removed

+++ Mewdeko.Core/Services/GreetSettingsService.cs
@@ -1,30 +1,27 @@
-﻿using Discord;
+﻿using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Discord;
 using Discord.WebSocket;
 using Mewdeko.Common;
 using Mewdeko.Common.Replacements;
 using Mewdeko.Core.Services.Database.Models;
 using Mewdeko.Extensions;
 using NLog;
-using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Mewdeko.Core.Services
 {
     public class GreetSettingsService : INService
     {
-        private readonly DbService _db;
-
-        public ConcurrentDictionary<ulong, GreetSettings> GuildConfigsCache { get; }
-        private readonly DiscordSocketClient _client;
-        private readonly Logger _log;
-
-        private GreetGrouper<IGuildUser> greets = new GreetGrouper<IGuildUser>();
-        private GreetGrouper<IGuildUser> byes = new GreetGrouper<IGuildUser>();
         private readonly BotSettingsService _bss;
-        public bool GroupGreets => _bss.Data.GroupGreets;
+        private readonly DiscordSocketClient _client;
+        private readonly DbService _db;
+        private readonly Logger _log;
+        private readonly GreetGrouper<IGuildUser> byes = new();
+
+        private readonly GreetGrouper<IGuildUser> greets = new();
 
         public GreetSettingsService(DiscordSocketClient client, Mewdeko bot, DbService db, BotSettingsService bss)
         {
@@ -43,6 +40,9 @@ namespace Mewdeko.Core.Services
             bot.JoinedGuild += Bot_JoinedGuild;
             _client.LeftGuild += _client_LeftGuild;
         }
+
+        public ConcurrentDictionary<ulong, GreetSettings> GuildConfigsCache { get; }
+        public bool GroupGreets => _bss.Data.GroupGreets;
 
         private Task _client_LeftGuild(SocketGuild arg)
         {
@@ -67,11 +67,13 @@ namespace Mewdeko.Core.Services
                     var conf = GetOrAddSettingsForGuild(user.GuildId);
 
                     if (!conf.SendChannelByeMessage) return;
-                    var channel = (await user.Guild.GetTextChannelsAsync().ConfigureAwait(false)).SingleOrDefault(c => c.Id == conf.ByeMessageChannelId);
+                    var channel =
+                        (await user.Guild.GetTextChannelsAsync().ConfigureAwait(false)).SingleOrDefault(c =>
+                            c.Id == conf.ByeMessageChannelId);
 
                     if (channel == null) //maybe warn the server owner that the channel is missing
                         return;
-                    
+
                     if (GroupGreets)
                     {
                         // if group is newly created, greet that user right away,
@@ -82,7 +84,7 @@ namespace Mewdeko.Core.Services
                             // greet single user
                             await ByeUsers(conf, channel, new[] {user});
                             var groupClear = false;
-                            while(!groupClear)
+                            while (!groupClear)
                             {
                                 await Task.Delay(5000).ConfigureAwait(false);
                                 groupClear = byes.ClearGroup(user.GuildId, 5, out var toBye);
@@ -120,12 +122,15 @@ namespace Mewdeko.Core.Services
         }
 
         private Task ByeUsers(GreetSettings conf, ITextChannel channel, IUser user)
-            => ByeUsers(conf, channel, new[] {user});
+        {
+            return ByeUsers(conf, channel, new[] {user});
+        }
+
         private async Task ByeUsers(GreetSettings conf, ITextChannel channel, IEnumerable<IUser> users)
         {
             if (!users.Any())
                 return;
-            
+
             var rep = new ReplacementBuilder()
                 .WithChannel(channel)
                 .WithClient(_client)
@@ -139,12 +144,12 @@ namespace Mewdeko.Core.Services
                 try
                 {
                     var toDelete = await channel.EmbedAsync(embedData).ConfigureAwait(false);
-                    if (conf.AutoDeleteByeMessagesTimer > 0)
-                    {
-                        toDelete.DeleteAfter(conf.AutoDeleteByeMessagesTimer);
-                    }
+                    if (conf.AutoDeleteByeMessagesTimer > 0) toDelete.DeleteAfter(conf.AutoDeleteByeMessagesTimer);
                 }
-                catch (Exception ex) { _log.Warn(ex); }
+                catch (Exception ex)
+                {
+                    _log.Warn(ex);
+                }
             }
             else
             {
@@ -154,23 +159,25 @@ namespace Mewdeko.Core.Services
                 try
                 {
                     var toDelete = await channel.SendMessageAsync(msg.SanitizeMentions()).ConfigureAwait(false);
-                    if (conf.AutoDeleteByeMessagesTimer > 0)
-                    {
-                        toDelete.DeleteAfter(conf.AutoDeleteByeMessagesTimer);
-                    }
+                    if (conf.AutoDeleteByeMessagesTimer > 0) toDelete.DeleteAfter(conf.AutoDeleteByeMessagesTimer);
                 }
-                catch (Exception ex) { _log.Warn(ex); }
+                catch (Exception ex)
+                {
+                    _log.Warn(ex);
+                }
             }
         }
-        
+
         private Task GreetUsers(GreetSettings conf, ITextChannel channel, IGuildUser user)
-            => GreetUsers(conf, channel, new[] {user});
-        
+        {
+            return GreetUsers(conf, channel, new[] {user});
+        }
+
         private async Task GreetUsers(GreetSettings conf, ITextChannel channel, IEnumerable<IGuildUser> users)
         {
             if (!users.Any())
                 return;
-            
+
             var rep = new ReplacementBuilder()
                 .WithChannel(channel)
                 .WithClient(_client)
@@ -184,10 +191,7 @@ namespace Mewdeko.Core.Services
                 try
                 {
                     var toDelete = await channel.EmbedAsync(embedData).ConfigureAwait(false);
-                    if (conf.AutoDeleteGreetMessagesTimer > 0)
-                    {
-                        toDelete.DeleteAfter(conf.AutoDeleteGreetMessagesTimer);
-                    }
+                    if (conf.AutoDeleteGreetMessagesTimer > 0) toDelete.DeleteAfter(conf.AutoDeleteGreetMessagesTimer);
                 }
                 catch (Exception ex)
                 {
@@ -198,29 +202,25 @@ namespace Mewdeko.Core.Services
             {
                 var msg = rep.Replace(conf.ChannelGreetMessageText);
                 if (!string.IsNullOrWhiteSpace(msg))
-                {
                     try
                     {
                         var toDelete = await channel.SendMessageAsync(msg.SanitizeMentions()).ConfigureAwait(false);
                         if (conf.AutoDeleteGreetMessagesTimer > 0)
-                        {
                             toDelete.DeleteAfter(conf.AutoDeleteGreetMessagesTimer);
-                        }
                     }
                     catch (Exception ex)
                     {
                         _log.Warn(ex);
                     }
-                }
             }
         }
 
         private async Task<bool> GreetDmUser(GreetSettings conf, IDMChannel channel, IGuildUser user)
         {
             var rep = new ReplacementBuilder()
-                .WithDefault(user, channel, (SocketGuild)user.Guild, _client)
+                .WithDefault(user, channel, (SocketGuild) user.Guild, _client)
                 .Build();
-            
+
             if (CREmbed.TryParse(conf.DmGreetMessageText, out var embedData))
             {
                 rep.Replace(embedData);
@@ -237,7 +237,6 @@ namespace Mewdeko.Core.Services
             {
                 var msg = rep.Replace(conf.DmGreetMessageText);
                 if (!string.IsNullOrWhiteSpace(msg))
-                {
                     try
                     {
                         await channel.SendConfirmAsync(msg).ConfigureAwait(false);
@@ -246,7 +245,6 @@ namespace Mewdeko.Core.Services
                     {
                         return false;
                     }
-                }
             }
 
             return true;
@@ -275,7 +273,7 @@ namespace Mewdeko.Core.Services
                                     // greet single user
                                     await GreetUsers(conf, channel, new[] {user});
                                     var groupClear = false;
-                                    while(!groupClear)
+                                    while (!groupClear)
                                     {
                                         await Task.Delay(5000).ConfigureAwait(false);
                                         groupClear = greets.ClearGroup(user.GuildId, 5, out var toGreet);
@@ -288,17 +286,13 @@ namespace Mewdeko.Core.Services
                                 await GreetUsers(conf, channel, new[] {user});
                             }
                         }
-
                     }
 
                     if (conf.SendDmGreetMessage)
                     {
                         var channel = await user.GetOrCreateDMChannelAsync().ConfigureAwait(false);
 
-                        if (channel != null)
-                        {
-                            await GreetDmUser(conf, channel, user);
-                        }
+                        if (channel != null) await GreetDmUser(conf, channel, user);
                     }
                 }
                 catch
@@ -339,9 +333,7 @@ namespace Mewdeko.Core.Services
                 settings.AutoDeleteByeMessagesTimer < 0 ||
                 settings.AutoDeleteGreetMessagesTimer > 600 ||
                 settings.AutoDeleteGreetMessagesTimer < 0)
-            {
                 return false;
-            }
 
             using (var uow = _db.GetDbContext())
             {
@@ -385,6 +377,7 @@ namespace Mewdeko.Core.Services
 
                 await uow.SaveChangesAsync();
             }
+
             return enabled;
         }
 
@@ -407,6 +400,7 @@ namespace Mewdeko.Core.Services
 
                 uow.SaveChanges();
             }
+
             return greetMsgEnabled;
         }
 
@@ -423,58 +417,9 @@ namespace Mewdeko.Core.Services
 
                 await uow.SaveChangesAsync();
             }
+
             return enabled;
         }
-
-        #region Get Enabled Status
-        public bool GetGreetDmEnabled(ulong guildId)
-        {
-            using (var uow = _db.GetDbContext())
-            {
-                var conf = uow.GuildConfigs.ForId(guildId, set => set);
-                return conf.SendDmGreetMessage;
-            }
-        }
-        
-        public bool GetGreetEnabled(ulong guildId)
-        {
-            using (var uow = _db.GetDbContext())
-            {
-                var conf = uow.GuildConfigs.ForId(guildId, set => set);
-                return conf.SendChannelGreetMessage;
-            }
-        }
-        
-        public bool GetByeEnabled(ulong guildId)
-        {
-            using (var uow = _db.GetDbContext())
-            {
-                var conf = uow.GuildConfigs.ForId(guildId, set => set);
-                return conf.SendChannelByeMessage;
-            }
-        }
-        #endregion
-        
-        #region Test Messages
-
-        public Task ByeTest(ITextChannel channel, IGuildUser user)
-        {
-            var conf = GetOrAddSettingsForGuild(user.GuildId);
-            return ByeUsers(conf, channel, user);
-        }
-        
-        public Task GreetTest(ITextChannel channel, IGuildUser user)
-        {
-            var conf = GetOrAddSettingsForGuild(user.GuildId);
-            return GreetUsers(conf, channel, user);
-        }
-        
-        public Task<bool> GreetDmTest(IDMChannel channel, IGuildUser user)
-        {
-            var conf = GetOrAddSettingsForGuild(user.GuildId);
-            return GreetDmUser(conf, channel, user);
-        }
-        #endregion
 
         public bool SetGreetDmMessage(ulong guildId, ref string message)
         {
@@ -495,6 +440,7 @@ namespace Mewdeko.Core.Services
 
                 uow.SaveChanges();
             }
+
             return greetMsgEnabled;
         }
 
@@ -512,6 +458,7 @@ namespace Mewdeko.Core.Services
 
                 await uow.SaveChangesAsync();
             }
+
             return enabled;
         }
 
@@ -534,6 +481,7 @@ namespace Mewdeko.Core.Services
 
                 uow.SaveChanges();
             }
+
             return byeMsgEnabled;
         }
 
@@ -570,6 +518,59 @@ namespace Mewdeko.Core.Services
                 await uow.SaveChangesAsync();
             }
         }
+
+        #region Get Enabled Status
+
+        public bool GetGreetDmEnabled(ulong guildId)
+        {
+            using (var uow = _db.GetDbContext())
+            {
+                var conf = uow.GuildConfigs.ForId(guildId, set => set);
+                return conf.SendDmGreetMessage;
+            }
+        }
+
+        public bool GetGreetEnabled(ulong guildId)
+        {
+            using (var uow = _db.GetDbContext())
+            {
+                var conf = uow.GuildConfigs.ForId(guildId, set => set);
+                return conf.SendChannelGreetMessage;
+            }
+        }
+
+        public bool GetByeEnabled(ulong guildId)
+        {
+            using (var uow = _db.GetDbContext())
+            {
+                var conf = uow.GuildConfigs.ForId(guildId, set => set);
+                return conf.SendChannelByeMessage;
+            }
+        }
+
+        #endregion
+
+        #region Test Messages
+
+        public Task ByeTest(ITextChannel channel, IGuildUser user)
+        {
+            var conf = GetOrAddSettingsForGuild(user.GuildId);
+            return ByeUsers(conf, channel, user);
+        }
+
+        public Task GreetTest(ITextChannel channel, IGuildUser user)
+        {
+            var conf = GetOrAddSettingsForGuild(user.GuildId);
+            return GreetUsers(conf, channel, user);
+        }
+
+        public Task<bool> GreetDmTest(IDMChannel channel, IGuildUser user)
+        {
+            var conf = GetOrAddSettingsForGuild(user.GuildId);
+            return GreetDmUser(conf, channel, user);
+        }
+
+        #endregion
     }
 
     public class GreetSettings
@@ -589,18 +590,21 @@ namespace Mewdeko.Core.Services
         public bool SendChannelByeMessage { get; set; }
         public string ChannelByeMessageText { get; set; }
 
-        public static GreetSettings Create(GuildConfig g) => new GreetSettings()
+        public static GreetSettings Create(GuildConfig g)
         {
-            AutoDeleteByeMessagesTimer = g.AutoDeleteByeMessagesTimer,
-            AutoDeleteGreetMessagesTimer = g.AutoDeleteGreetMessagesTimer,
-            GreetMessageChannelId = g.GreetMessageChannelId,
-            ByeMessageChannelId = g.ByeMessageChannelId,
-            SendDmGreetMessage = g.SendDmGreetMessage,
-            DmGreetMessageText = g.DmGreetMessageText,
-            SendChannelGreetMessage = g.SendChannelGreetMessage,
-            ChannelGreetMessageText = g.ChannelGreetMessageText,
-            SendChannelByeMessage = g.SendChannelByeMessage,
-            ChannelByeMessageText = g.ChannelByeMessageText,
-        };
+            return new()
+            {
+                AutoDeleteByeMessagesTimer = g.AutoDeleteByeMessagesTimer,
+                AutoDeleteGreetMessagesTimer = g.AutoDeleteGreetMessagesTimer,
+                GreetMessageChannelId = g.GreetMessageChannelId,
+                ByeMessageChannelId = g.ByeMessageChannelId,
+                SendDmGreetMessage = g.SendDmGreetMessage,
+                DmGreetMessageText = g.DmGreetMessageText,
+                SendChannelGreetMessage = g.SendChannelGreetMessage,
+                ChannelGreetMessageText = g.ChannelGreetMessageText,
+                SendChannelByeMessage = g.SendChannelByeMessage,
+                ChannelByeMessageText = g.ChannelByeMessageText
+            };
+        }
     }
 }
